refactor(package): 데이터 계층의 모델 패키지 상세화
