+++ app/src/main/java/woowacourse/shopping/data/cart/model/dto/response/CartProductResponse.kt
@@ -1,0 +1,9 @@
+package woowacourse.shopping.data.cart.model.dto.response
+
+import woowacourse.shopping.data.product.model.dto.response.ProductResponse
+
+data class CartProductResponse(
+    val id: Long,
+    val product: ProductResponse,
+    val quantity: Int,
+)
