+++ app/src/main/java/woowacourse/shopping/data/order/model/Mapper.kt
@@ -3,6 +3,9 @@ package woowacourse.shopping.data.order.model
 import com.example.domain.FixedDiscountPolicies
 import com.example.domain.FixedDiscountPolicy
 import com.example.domain.order.OrderSummary
+import woowacourse.shopping.data.order.model.dto.response.FixedDiscountPoliciesResponse
+import woowacourse.shopping.data.order.model.dto.response.FixedDiscountPolicyResponse
+import woowacourse.shopping.data.order.model.dto.response.OrderSummaryResponse
 
 fun FixedDiscountPolicyResponse.toDomain(): FixedDiscountPolicy {
     return FixedDiscountPolicy(
