+++ app/src/main/java/woowacourse/shopping/data/order/model/dto/request/OrderRequest.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.order.model
+package woowacourse.shopping.data.order.model.dto.request
 
 data class OrderRequest(
     val cartItemIds: List<Long>,
