+++ app/src/main/java/woowacourse/shopping/data/order/model/dto/response/FixedDiscountPoliciesResponse.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.order.model
+package woowacourse.shopping.data.order.model.dto.response
 
 data class FixedDiscountPoliciesResponse(
     val fixedDiscountPolicies: List<FixedDiscountPolicyResponse>
