+++ app/src/main/java/woowacourse/shopping/data/order/model/dto/response/OrderSummaryResponse.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.order.model
+package woowacourse.shopping.data.order.model.dto.response
 
 data class OrderSummaryResponse(
     val id: Long,
