--- app/src/main/java/woowacourse/shopping/data/product/model/ProductsResponse.kt
@@ -1,6 +1,0 @@
-package woowacourse.shopping.data.product.model
-
-data class ProductsResponse(
-    val products: ProductsDto,
-    val pagination: PaginationDto,
-)
