+++ app/src/main/java/woowacourse/shopping/data/product/model/dto/ProductsDto.kt
@@ -1,0 +1,3 @@
+package woowacourse.shopping.data.product.model.dto
+
+class ProductsDto : ArrayList<ProductDto>()
